requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
dateutils==0.6.12
Django==4.2.4
idna==3.4
nameparser==1.1.2
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
ujson==5.8.0
urllib3==2.0.4
